+++ extensions/time-tracking/src/view-project-timers.tsx
@@ -11,6 +11,7 @@ import {
   List,
   showToast,
   Form,
+  getPreferenceValues,
 } from "@raycast/api";
 
 import { useEffect, useState } from "react";
@@ -24,54 +25,62 @@ import {
   Timer,
   TimerList,
 } from "./Timers";
+import { useForm } from "@raycast/utils";
 
 function EditForm(props: { timer: Timer; onUpdate: (start: Date, end: Date, name: string, tag: string) => void }) {
-  const [error, setError] = useState("");
+  const { default_tag } = getPreferenceValues<Preferences.StartTimer>();
+  type FormValues = {
+    id: string;
+    name: string;
+    tag: string;
+    start: Date | null;
+    end: Date | null;
+  };
+  const { itemProps, handleSubmit, values, setValidationError } = useForm<FormValues>({
+    onSubmit(values) {
+      const { start, end, name, tag } = values;
+      if (!start) {
+        setValidationError("start", "The item is required");
+        return;
+      }
+      if (!end) {
+        setValidationError("end", "The item is required");
+        return;
+      }
+      props.onUpdate(start, end, name, tag);
+    },
+    initialValues: {
+      id: props.timer.id,
+      name: props.timer.name || "",
+      tag: props.timer.tag || "",
+      start: new Date(props.timer.start),
+      end: new Date(props.timer.end!),
+    },
+    validation: {
+      start(value) {
+        if (!value) return "The item is required";
+        if (value > new Date()) return "Start Date must be a date in the past";
+        if (values.end && values.end <= value) return "Start Date must be before End Date";
+      },
+      end(value) {
+        if (!value) return "The item is required";
+        if (value > new Date()) return "End Date must be a date in the past";
+        if (values.start && values.start >= value) return "End Date must be after Start Date";
+      },
+    },
+  });
   return (
     <Form
       actions={
         <ActionPanel>
-          <Action.SubmitForm
-          icon={Icon.EditShape}
-            title="Submit"
-            onSubmit={(input) => {
-              const name: string = input.name;
-              const start: Date = input["start-date"];
-              const end: Date = input["end-date"];
-              const tag: string = input.tag;
-              if (start >= end) {
-                setError("End Date must be after Start Date");
-                return false;
-              }
-              if (end > new Date()) {
-                setError("End Date must be a date in the past");
-                return false;
-              }
-              props.onUpdate(start, end, name, tag);
-            }}
-          />
+          <Action.SubmitForm icon={Icon.EditShape} title="Submit" onSubmit={handleSubmit} />
         </ActionPanel>
       }
     >
-      <Form.TextField title="Name" id="name" defaultValue={props.timer.name || undefined} placeholder="Unnamed timer" />
-      <Form.DatePicker
-        title="Start Date"
-        id={"start-date"}
-        defaultValue={new Date(props.timer.start)}
-        // This is important to clear the state of the form. Without it, subsequent submits
-        // will send stale values.
-        onChange={() => setError("")}
-      />
-      <Form.DatePicker
-        error={error}
-        title="End Date"
-        id={"end-date"}
-        defaultValue={new Date(props.timer.end!)}
-        // This is important to clear the state of the form. Without it, subsequent submits
-        // will send stale values.
-        onChange={() => setError("")}
-      />
-      <Form.TextField title="Tag" id="tag" defaultValue={props.timer.tag || undefined} placeholder="Tag" />
+      <Form.TextField title="Name" placeholder="Unnamed timer" {...itemProps.name} />
+      <Form.DatePicker title="Start Date" {...itemProps.start} />
+      <Form.DatePicker title="End Date" {...itemProps.end} />
+      <Form.TextField title="Tag" placeholder={default_tag || "Tag"} {...itemProps.tag} />
     </Form>
   );
 }
@@ -134,12 +143,10 @@ export default function Command() {
               : new Date(timer.start).toLocaleString()
           }
           accessories={[
+            { tag: timer.tag },
             {
               text: (timer.end ? "✅ " : "⏳ ") + formatDuration(getDuration(timer)),
             },
-            {
-              tag: timer.tag
-            }
           ]}
           actions={
             <ActionPanel>
@@ -193,7 +200,7 @@ export default function Command() {
                       message: "This action cannot be undone.",
                       icon: { source: Icon.Trash, tintColor: Color.Red },
                       primaryAction: {
-                        style: Alert.ActionStyle.Default,
+                        style: Alert.ActionStyle.Destructive,
                         title: "Delete timer",
                       },
                       rememberUserChoice: true,